+++ app/db/models/property_amenity.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey
+from sqlalchemy import Column, BigInteger, ForeignKey
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import relationship
 from app.db.sessions.session import Base
@@ -6,7 +6,7 @@ from app.db.sessions.session import Base
 class PropertyAmenity(Base):
     __tablename__ = "property_amenities"
 
-    property_id = Column(Integer, ForeignKey("properties.id"), primary_key=True)
+    property_id = Column(BigInteger, ForeignKey("properties.id"), primary_key=True)
     amenity_id = Column(UUID(as_uuid=True), ForeignKey("amenities.id"), primary_key=True)
 
     # Relationships
